Implement Godoc parsing and embedding.
This implements a system to parse the doc strings from packages / types and embed them inside the resulting build.

This improves the workflow for documenting types for consumption in Godot.

The package docs are parsed at build time and serialized to a gob written to the projects working directory.

An init file is also written to the working directory to allow embedding the gob inside the final extension; this makes a call to pass the binary gob data to the docs package for decoding at runtime.

Updates to the register_class file read from the registered docs and pass the docstrings to Godot